Updated script to ignore svn, git, and vim junk.
This script was ugly to begin with, and I just made it worse. I think I
need to simplify it, but it's not a high priority. Still, I think it
would be better just to call 'zip -0 -r tably.xpi blah/*' because that
would automatically ignore dot files: */.svn/ and */.git/ and .*.swp